Move Meta in components
